requirements.txt
Django==4.2.4
gunicorn==25.1.0
dj-database-url==2.2.0
psycopg2-binary==2.9.10
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.4.0